requirements.txt
apache-flink==1.18.1
apache-beam==2.48.0
redis==5.0.1
kafka-python==2.0.2
pytest==8.0.0

